# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "灵嘚中间件（LingDei-Middleware）是一个基于 Fiber 的 RESTful API 服务，用于提供灵嘚（LingDei）的后端服务。\n注意，有 🦸 标识的接口需要管理员权限才能访问。",
        "title": "灵嘚中间件",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "127.0.0.1:9000",
    "basePath": "/",
    "paths": {
...
        "/fan/count": {
            "get": {
                "description": "获取粉丝数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "粉丝管理"
                ],
                "summary": "获取粉丝数量",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户UUID",
                        "name": "follow_uuid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowCountResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/fan/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取我的粉丝列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "粉丝管理"
                ],
                "summary": "获取我的粉丝列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/follow/add": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "关注用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注管理"
                ],
                "summary": "关注用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "被关注用户UUID",
                        "name": "follow_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/follow/check": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取关注状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注管理"
                ],
                "summary": "获取关注状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "被关注用户UUID",
                        "name": "follow_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowStatusResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.FollowStatusResp"
                        }
                    }
                }
            }
        },
        "/follow/count": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取已关注up主的数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注管理"
                ],
                "summary": "获取已关注up主的数量",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowCountResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/follow/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "取消关注用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注管理"
                ],
                "summary": "取消关注用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "被关注用户UUID",
                        "name": "follow_uuid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
        "/follow/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取我的关注列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "关注管理"
                ],
                "summary": "获取我的关注列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页数",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FollowListResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.OperationResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Follow": {
            "type": "object",
            "required": [
                "follow_uuid",
                "user_uuid",
                "uuid"
            ],
            "properties": {
                "follow_uuid": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                },
                "user_uuid": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.FollowCountResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "model.FollowListResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "follow_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Follow"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.FollowStatusResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "model.OperationResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.Profile": {
            "type": "object",
            "required": [
                "id",
                "nickname"
            ],
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string",
                    "maxLength": 25,
                    "minLength": 2
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "required": [
                "id",
                "username"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "profile": {
                    "description": "Profile",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Profile"
                        }
                    ]
                },
                "role": {
                    "type": "integer"
                },
                "username": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 3
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}